requirements.txt
aiogram~=2.12
arrow==1.2.0
environs~=9.3.2
nltk~=3.6.2
python-miio~=0.5.6
PyAudio~=0.2.11
SQLAlchemy~=1.4.22
numpy==1.21.2
PyYAML~=5.4.1
wheel==0.37.0
fuzzywuzzy~=0.18.0
gTTS~=2.2.3
playsound==1.3.0
SpeechRecognition~=3.8.1
vosk~=0.3.31
scipy~=1.7.1
pyttsx3~=2.90
googletrans~=3.0.0
django-googlesearch~=0.2
pyowm~=3.2.0
termcolor~=1.1.0
python-dotenv~=0.19.1
scikit-learn~=1.0
pika~=1.2.0